+++ src/us/mn/state/dot/tms/client/incident/IncidentLoc.java
@@ -119,22 +119,22 @@ public class IncidentLoc implements GeoLoc {
 	}
 
 	/** Set the latitude */
-	public void setLat(Float lt) {
+	public void setLat(Double lt) {
 		// part of GeoLoc interface
 	}
 
 	/** Get the latitude */
-	public Float getLat() {
+	public Double getLat() {
 		return incident.getLat();
 	}
 
 	/** Set the longitude */
-	public void setLon(Float ln) {
+	public void setLon(Double ln) {
 		// part of GeoLoc interface
 	}
 
 	/** Get the longitude */
-	public Float getLon() {
+	public Double getLon() {
 		return incident.getLon();
 	}
